requirements.txt
langchain>=0.3.0
langchain-google-genai>=3.0.0
google-generativeai
python-dotenv